requirements.txt
# requirements.txt
werkzeug===2.2.2
Flask
flask-login
flask-wtf
Flask-SQLAlchemy
Flask-Bcrypt
python-dotenv
dash
dash-bootstrap-components
pandas